+++ Code/Scenes/LevelScene.cpp
@@ -91,7 +91,7 @@ void LevelScene::Load()
 	float	thickWall			=	3.f;
 
 	//map entities
-	for (int x = map.playMap.mapNodes[0].size() - 1; x >= 0 ; x--) { for (int y = map.playMap.mapNodes.size() - 1; y >= 0; y--) {
+	for (int x = (int)map.playMap.mapNodes[0].size() - 1; x >= 0 ; x--) { for (int y = (int)map.playMap.mapNodes.size() - 1; y >= 0; y--) {
 		auto mapNode = std::make_unique<Entity>("MapNode_" + std::to_string(x) + "_" + std::to_string(y));
 		auto mapFog = std::make_unique<Entity>("MapFog_" + std::to_string(x) + "_" + std::to_string(y));
 		pos = { minimapOffX + (float)x * NODE_SIZE, minimapOffY - (float)y * NODE_SIZE };
@@ -215,7 +215,7 @@ void LevelScene::Load()
 	scale = { 100.f, 100.f };
 	e->addComponent<Transform2D>(pos, scale, 0.f);
 	e->addComponent<Mesh>("Box", "player_sprite", Color(255, 255, 255, 1.f), 100, MeshType::BOX_T);
-	e->addComponent<BoxCollider2D>(scale.x / 1.5, scale.y / 1.5);
+	e->addComponent<BoxCollider2D>(scale.x / 1.5f, scale.y / 1.5f);
 	e->addComponent<Player>();
 	//e->addComponent<Bounce>(0.f, 2.f, 0.1f, 0.07f);
 	enSystem->rootEntity->transform->AddChild(e->transform);
@@ -229,7 +229,7 @@ void LevelScene::Load()
 	pos = { 0.f, (float)AEGfxGetWindowHeight() / 2 - scale.y/2};
 	n_path->addComponent<Transform2D>(pos, scale, 0.f);
 	//n_path->addComponent<Mesh>("Box", Color(255, 255, 255, 0.3), 100, MeshType::BOX_B);
-	n_path->addComponent<BoxCollider2D>(scale.x/ 1.5, scale.y/ 1.5);
+	n_path->addComponent<BoxCollider2D>(scale.x/ 1.5f, scale.y/ 1.5f);
 	n_path->addComponent<SceneEdge>();
 	enSystem->rootEntity->transform->AddChild(n_path->transform);
 	
@@ -248,7 +248,7 @@ void LevelScene::Load()
 	pos = { (float)AEGfxGetWindowWidth()/2 - scale.x/2, 0.f };
 	e_path->addComponent<Transform2D>(pos, scale, 0.f);
 	//e_path->addComponent<Mesh>("Box", Color(255, 255, 255, 0.3), 100, MeshType::BOX_B);
-	e_path->addComponent<BoxCollider2D>(scale.x / 1.5, scale.y / 1.5);
+	e_path->addComponent<BoxCollider2D>(scale.x / 1.5f, scale.y / 1.5f);
 	e_path->addComponent<SceneEdge>();
 	enSystem->rootEntity->transform->AddChild(e_path->transform);
 
@@ -264,9 +264,9 @@ void LevelScene::Load()
 
 	auto s_path = std::make_unique<Entity>("S_Path");
 	scale = { (float)AEGfxGetWindowWidth(), collidersize };
-	pos = { 0.f, scale.y/2 - (float)AEGfxGetWindowHeight()/2};
+	pos = { 0.f, scale.y/2.f - (float)AEGfxGetWindowHeight()/2.f};
 	s_path->addComponent<Transform2D>(pos, scale, 0.f);
-	s_path->addComponent<BoxCollider2D>(scale.x / 1.5, scale.y / 1.5);
+	s_path->addComponent<BoxCollider2D>(scale.x / 1.5f, scale.y / 1.5f);
 	s_path->addComponent<SceneEdge>();
 	enSystem->rootEntity->transform->AddChild(s_path->transform);
 
@@ -284,7 +284,7 @@ void LevelScene::Load()
 	scale = { collidersize, (float)AEGfxGetWindowHeight() };
 	pos = { scale.x/2 - (float)AEGfxGetWindowWidth() / 2, 0.f };
 	w_path->addComponent<Transform2D>(pos, scale, 0.f);
-	w_path->addComponent<BoxCollider2D>(scale.x / 1.5, scale.y / 1.5);
+	w_path->addComponent<BoxCollider2D>(scale.x / 1.5f, scale.y / 1.5f);
 	w_path->addComponent<SceneEdge>();
 	enSystem->rootEntity->transform->AddChild(w_path->transform);
